% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_backups}
\alias{list_backups}
\title{List Available Backups}
\usage{
list_backups(backup_dir = ".", type = c("github", "cran", "all"))
}
\arguments{
\item{backup_dir}{Backup directory to scan}

\item{type}{Type of backup ("github", "cran", or "all")}
}
\value{
Data frame with backup information
}
\description{
List all available package backups in the specified directory.
}
\examples{
\dontrun{
# List all GitHub backups
github_backups <- list_backups("github_backup", type = "github")

# List all backups
all_backups <- list_backups(type = "all")
}

}
